feat: config layer
