+++ src/anonymizer/config/model_configs/entity_detection.yaml
@@ -1,0 +1,5 @@
+selected_models:
+  entity_detector: gliner-pii-detector
+  entity_validator: build-gpt-oss-120b
+  entity_augmenter: build-gpt-oss-120b
+  latent_detector: nemotron-30b-thinking
