+++ src/anonymizer/config/model_configs/replace_workflow.yaml
@@ -1,0 +1,2 @@
+selected_models:
+  replacer: build-gpt-oss-120b
